fix: resolve build errors blocking Vercel deployment
- Move stripe webhook idempotency logic to separate module (route files
  can only export HTTP methods + config)
- Fix Stripe Invoice.subscription type error (use any cast)
- Fix vitest.config.ts environmentMatchGlobs not supported (use jsdom
  for all tests)
